requirements.txt
pyrogram>=1.0.0
telegraph==1.4.1
tgcrypto
python-dotenv